small_website/requirements.txt
absl-py==0.7.1
aiohttp==3.5.4
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
APScheduler==3.5.3
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
billiard==3.3.0.23
bleach==3.1.0
blinker==1.4
brotlipy==0.7.0
bs4==0.0.1
cchardet==2.1.4
celery==3.1.26.post2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.6.1
colorama==0.4.0
constantly==15.1.0
crypto==1.4.1
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
dask==0.20.1
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.5
django-apscheduler==0.2.13
django-celery==3.2.2
django-cors-headers==2.4.0
docopt==0.6.2
easydict==1.9
entrypoints==0.3
fake-useragent==0.1.11
Flask==1.0.2
Flask-Login==0.4.1
fonttools==3.42.0
furl==2.0.0
gast==0.2.2
gerapy==0.8.5
gevent==1.3.7
greenlet==0.4.15
grpcio==1.20.1
h11==0.7.0
h2==3.0.1
h5py==2.9.0
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.8
idna-ssl==1.1.0
incremental==17.5.0
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsmin==2.2.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kaitaistruct==0.8
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
kombu==3.0.37
ldap3==2.5.1
lxml==4.2.5
Markdown==3.1
MarkupSafe==1.1.0
matplotlib==3.0.1
mistune==0.8.4
mitmproxy==4.0.4
mock==3.0.4
multidict==4.5.2
Naked==0.1.31
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.16.2
opencv-python==4.0.0.21
orderedmultidict==1.0
pandas==0.23.4
pandocfilters==1.4.2
paramiko==2.4.2
parse==1.12.0
parsel==1.5.1
parso==0.3.1
passlib==1.7.1
pickleshare==0.7.5
pika==0.12.0
Pillow==5.3.0
pipenv==2018.11.26
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.7.1
py==1.8.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAutoGUI==0.9.38
pycparser==2.19
pycrypto==2.6.1
pycryptodomex==3.8.1
pycurl==7.43.0.2
PyDispatcher==2.0.5
pydivert==2.1.0
pyee==6.0.0
PyExecJS==1.5.1
Pygments==2.2.0
PyHamcrest==1.9.0
pymongo==3.7.2
PyMsgBox==1.0.6
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
pyppeteer==0.0.25
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyquery==1.4.0
PyScreeze==0.1.19
pyspider==0.3.10
pytesseract==0.2.6
python-dateutil==2.7.5
python-decouple==3.1
python-scrapyd-api==2.1.2
PyTweening==1.0.3
pytz==2018.7
PyWavelets==1.0.1
pywin32==223
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
queuelib==1.5.0
redis==3.0.1
requests==2.21.0
requests-html==0.10.0
requests-toolbelt==0.8.0
retry==0.9.2
retrying==1.3.3
ruamel.yaml==0.15.77
scikit-image==0.14.1
scipy==1.1.0
Scrapy==1.6.0
scrapy-redis==0.6.8
scrapy-splash==0.7.2
selenium==3.141.0
Send2Trash==1.5.0
service-identity==17.0.0
shellescape==3.4.1
six==1.11.0
sortedcontainers==2.0.5
soupsieve==1.8
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
threadpool==1.3.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
twisted==18.9.0
typing-extensions==3.7.2
tzlocal==2.0.0b1
u-msgpack-python==2.5.1
udunits2===2-2.2.27.6
urllib3==1.24.1
urwid==2.0.1
vine==1.2.0
virtualenv==16.4.3
virtualenv-clone==0.5.1
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==7.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
WsgiDAV==3.0.0
wsproto==0.11.0
xlrd==1.1.0
XlsxWriter==1.1.2
yarl==1.3.0
zope.interface==4.6.0
